+++ mkdocs_pdf_generate/utils.py
@@ -1,14 +1,12 @@
 import os
 import re
-from typing import Optional, Union, Dict
+from typing import Dict, Optional, Union
 
 from bs4 import BeautifulSoup, PageElement
 
 
 def get_pdf_metadata(metadata):
-    pdf_meta = (
-        metadata.get("pdf") if "pdf" in metadata and metadata["pdf"] is not None else {}
-    )
+    pdf_meta = metadata.get("pdf") if "pdf" in metadata and metadata["pdf"] is not None else {}
     return pdf_meta
 
 
@@ -52,26 +50,18 @@ def secure_filename(filename):
     for sep in os.path.sep, os.path.altsep:
         if sep:
             filename = filename.replace(sep, " ")
-    filename = str(_filename_ascii_strip_re.sub("", "_".join(filename.split()))).strip(
-        "._"
-    )
+    filename = str(_filename_ascii_strip_re.sub("", "_".join(filename.split()))).strip("._")
 
     # on nt a couple of special files are present in each folder.  We
     # have to ensure that the target file is not such a filename.  In
     # this case we prepend an underline
-    if (
-        os.name == "nt"
-        and filename
-        and filename.split(".")[0].upper() in _windows_device_files
-    ):
+    if os.name == "nt" and filename and filename.split(".")[0].upper() in _windows_device_files:
         filename = "_" + filename
 
     return filename
 
 
-def h1_title_tag(
-    content: Union[str, PageElement], page_metadata: Dict
-) -> Optional[str]:
+def h1_title_tag(content: Union[str, PageElement], page_metadata: Dict) -> Optional[str]:
     soup = content
     if isinstance(soup, str):
         soup = BeautifulSoup(soup, "html5lib")